addVectorTiles

Add vector tiles layer to leaflet map using the Leaflet.VectorGrid API reference.

addVectorTiles(
  map,
  server,
  layer,
  filter = "",
  layerId,
  style = list(fill = TRUE, fillColor = "purple", fillOpacity = 0.2, color = "purple",
    opacity = 0.7, weight = 0.1, radius = 0.8),
  styleHighlight = list(fill = TRUE, fillColor = "yellow", fillOpacity = 0.2, color =
    "red", opacity = 0.7, weight = 0.1, radius = 0.8),
  group = "",
  pane = "overlayPane",
  attribution = NULL,
  debug = FALSE
)

Arguments

map

map object from leaflet::leaflet()

server

URL to tile server

layer

layer name

filter

filter string using Common Query Language (CQL) syntax. Defaults to an empty string "".

layerId

column name to uniquely identify each feature in layer

style

style as a list object

styleHighlight

style as a list object for clicked feature

group

group

pane

pane in leaflet map

attribution

attribution for layer

debug

show the JavaScript code for the map |> onRender(js_lyr)

Value

map object from leaflet::leaflet() with vector tile layer added

Examples

library(leaflet)

leaflet() |>
  addTiles() |>
  setView(-122.4, 33.8, 4) |>
  addVectorTiles(
    server  = "https://tile.calcofi.io",
    layer   = "public.stations",
    layerId = "stationid")